sync

+++ backend/src/main/java/waa/miu/finalproject/service/CommentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CommentService {
     public void save(InputCommentDto comment);
-    public SimpleCommentDto getComment(int id);
+    public SimpleCommentDto getComment(long id);
     public List<SimpleCommentDto> getPaginationComments(int postId);
     public List<SimpleCommentDto> getComments();
 }
